+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+##### 1.9.0 (2017-04-09)
+
+* Add `title` to `RedditButton` and `LinkedInButton`.
+
+##### 1.8.0 (2017-03-21)
+
+* Add default traget blank to EmailButton.
+* Add `title` and `message` parameter to VK button.
+
 ##### 1.7.1 (2017-01-20)
 
 * Use `media` as `picture` in FacebookButton.